+++ src/console.rs
