+++ src/lib.rs
@@ -1,4 +1,6 @@
 pub mod args;
+pub mod command_manager;
+pub mod console;
 pub mod demo;
 pub mod events;
 pub mod gamefinder;
@@ -9,8 +11,7 @@ pub mod player_records;
 pub mod server;
 pub mod settings;
 pub mod state;
-pub mod steamapi;
-// pub mod web;
+pub mod web;
 
 pub use clap;
 pub use rcon;
